Fix the type of kTLEBSize
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@2927 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ llvm/tsan_rtl/tsan_rtl.cc
@@ -38,7 +38,7 @@
 #endif
 
 extern bool global_ignore;
-static const int kTLEBSize = 2000;
+static const size_t kTLEBSize = 2000;
 
 struct ThreadInfo {
   tid_t tid;
